requirements.txt
aiogram>=3.0,<4.0
python-dotenv>=1.0.0
httpx>=0.25.0
trafilatura>=1.6.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.24.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.0
